+++ VRDR.Tests/fixtures/json/CauseOfDeathCodingUpdateMessage.json
@@ -9,7 +9,7 @@
             "resource": {
                 "resourceType": "MessageHeader",
                 "id": "a3a1ff4e-fc50-47eb-b3af-442e5fceadd1",
-                "eventUri": "http://nchs.cdc.gov/vrdr_demographics_coding",
+                "eventUri": "http://nchs.cdc.gov/vrdr_causeofdeath_coding_update",
                 "destination": [{"endpoint": "destination"}],
                 "source": {"endpoint": "http://nchs.cdc.gov/vrdr_submission"},
                 "focus": [{"reference": "urn:uuid:041b337f-8a85-4e83-9c37-b50610b1be88"}]
@@ -36,93 +36,6 @@
                     {
                         "name": "jurisdiction_id",
                         "valueString": "MA"
-                    },
-                    {
-                        "name": "cs",
-                        "valueCoding": {
-                            "system": "https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Software/MICAR/Data_Entry_Software/ACME_TRANSAX/Documentation/auser.pdf",
-                            "code": "8"
-                        }
-                    },
-                    {
-                        "name": "ship",
-                        "valueString": "B202101"
-                    },
-                    {
-                        "name": "rec_mo",
-                        "valueUnsignedInt": 1
-                    },
-                    {
-                        "name": "rec_dy",
-                        "valueUnsignedInt": 8
-                    },
-                    {
-                        "name": "rec_yr",
-                        "valueUnsignedInt": 2021
-                    },
-                    {
-                        "name": "manner",
-                        "valueString": "Accident"
-                    },
-                    {
-                        "name": "int_rej",
-                        "valueCoding": {
-                            "system": "https://www.cdc.gov/nchs/data/dvs/2b_2017.pdf",
-                            "code": "5"
-                        }
-                    },
-                    {
-                        "name": "sys_rej",
-                        "valueString": "ACMEReject"
-                    },
-                    {
-                        "name": "injpl",
-                        "valueString": "Home"
-                    },
-                    {
-                        "name": "ethnicity",
-                        "part": [
-                            {
-                                "name": "DETHNICE",
-                                "valueCoding": {
-                                    "system": "https://www.cdc.gov/nchs/data/dvs/HispanicCodeTitles.pdf",
-                                    "code": "123"
-                                }
-                            },
-                            {
-                                "name": "DETHNIC5C",
-                                "valueCoding": {
-                                    "system": "https://www.cdc.gov/nchs/data/dvs/HispanicCodeTitles.pdf",
-                                    "code": "456"
-                                }
-                            }
-                        ]
-                    },
-                    {
-                        "name": "race",
-                        "part": [
-                            {
-                                "name": "RACE1E",
-                                "valueCoding": {
-                                    "system": "https://www.cdc.gov/nchs/data/dvs/RaceCodeList.pdf",
-                                    "code": "foo"
-                                }
-                            },
-                            {
-                                "name": "RACE17C",
-                                "valueCoding": {
-                                    "system": "https://www.cdc.gov/nchs/data/dvs/RaceCodeList.pdf",
-                                    "code": "bar"
-                                }
-                            },
-                            {
-                                "name": "RACEBRG",
-                                "valueCoding": {
-                                    "system": "https://www.cdc.gov/nchs/data/dvs/Multiple_race_documentation_5-10-04.pdf",
-                                    "code": "baz"
-                                }
-                            }
-                        ]
                     },
                     {
                         "name": "underlying_cause_of_death",
@@ -169,7 +82,7 @@
                         "part": [
                             {
                                 "name": "lineNumber",
-                                "valueId": "abcde"
+                                "valueId": "1"
                             },
                             {
                                 "name": "coding",
@@ -192,7 +105,7 @@
                         "part": [
                             {
                                 "name": "lineNumber",
-                                "valueId": "xyzzy"
+                                "valueId": "2"
                             },
                             {
                                 "name": "coding",
